ref(callkit): move feature to call-integration directory
