+++ react/features/mobile/call-integration/CallKitIcon.png
