+++ react/features/mobile/call-integration/ConnectionService.js
