+++ react/features/mobile/call-integration/index.js
